Исправлена структура проекта
