+++ README.md
@@ -1,3 +1,21 @@
+# Тестируем классы интернет-магазина
+Вам нужно реализовать и протестировать классы интернет-магазина.
+Все места, которые нужно дописать как в тестах, так и классах, отмечены `# TODO`.
+
+При реализации обращайте внимание на типизацию аргументов методов и возвращаемых значений.
+Так же обратите внимание на организацию тестов в файле с тестами:
+- Тесты сгруппированы по классу, который они тестируют.
+- Каждый тест называется именем соответствующего ему метода.
+
+Вы можете начать как с реализации классов, так и с тестов.
+
+
+# Дополнительные вопросы:
+1. С чего проще начать выполнение домашнего задания: с тестов или с реализации классов?
+2. Почему для хранения товаров в корзине используется словарь, а не список?
+3. Зачем нужен __hash__ в классе Product? Изучите этот метод и объясните, почему он нужен.
+
+------------
 
 # Инкапсуляция
 Самодокументируемость кода, а так же соблюдение принципов вроде DRY и KISS
