+++ models/cart.py
@@ -1,42 +1,5 @@
-class Product:
-    """
-    Класс продукта
-    """
-    name: str
-    price: float
-    description: str
-    quantity: int
-
-    def __init__(self, name, price, description, quantity):
-        self.name = name
-        self.price = price
-        self.description = description
-        self.quantity = quantity
-
-    def check_quantity(self, quantity) -> bool:
-        """
-        TODO Верните True если количество продукта больше или равно запрашиваемому
-            и False в обратном случае
-        """
-        return True if (self.quantity >= quantity >= 0) else False
-
-
-    def buy(self, quantity):
-        """
-        TODO реализуйте метод покупки
-            Проверьте количество продукта используя метод check_quantity
-            Если продуктов не хватает, то выбросите исключение ValueError
-        """
-        if self.check_quantity(quantity):
-            self.quantity -= ValueChecker.check(quantity)
-            return quantity
-        else:
-            raise ValueError(f"Превышено максимально-возможное значение. Max - {self.quantity}, buy - {quantity}")
-
-
-    def __hash__(self):
-        """Необходимо для использования в качестве ключа в словарях"""
-        return hash(self.name + self.description)
+from models.product import Product
+from utils.value_cheker import ValueChecker
 
 
 class Cart:
@@ -102,12 +65,3 @@ class Cart:
         # товара на складе
         for position, count in self.products.items():
             position.buy(count)
-
-
-class ValueChecker:
-    @classmethod
-    def check(cls, value):
-        if value >= 0:
-            return value
-        else:
-            raise ValueError("Отрицательное значение параметра")
