+++ tests/test_cart.py
@@ -1,0 +1,10 @@
+class TestCart:
+    """
+    TODO Напишите тесты на методы класса Cart
+        На каждый метод у вас должен получиться отдельный тест
+        На некоторые методы у вас может быть несколько тестов.
+        Например, негативные тесты, ожидающие ошибку (используйте pytest.raises, чтобы проверить это)
+    """
+
+    def test_all(self, cart, product):
+        print(1)
