+++ utils/value_cheker.py
@@ -1,0 +1,8 @@
+class ValueChecker:
+
+    @classmethod
+    def check(cls, value):
+        if value >= 0:
+            return value
+        else:
+            raise ValueError("Отрицательное значение параметра")
